refactor: 로그인 경로 변경, AdminController -> ViewController로 변경
